* commit '01e07e1fe33ed868bbb2579334445803e253bc28':
  Add copyright to all relevant files.
  Attempt to ignore NotImplementedError that can't be tested (reference: https://github.com/codecov/codecov-python/issues/152#issuecomment-396881072).
  Add test of __repr__ for HSL and RGB color classes.
  Correct test input type for HSL from int to float.

+++ test/model/hsl/hsl_exception_handling_test.py
@@ -1,3 +1,5 @@
+# Copyright 2022 – present, Jakob Bagterp. BSD 3-Clause license and refer to LICENSE file.
+
 from contextlib import nullcontext as does_not_raise
 from typing import Any
 
@@ -21,12 +23,12 @@ MOCK_DATASET = [
 
 
 @pytest.mark.parametrize("hsl, expectation", MOCK_DATASET)
-def test_exception_handling_of_color_hsl(hsl: tuple[int, int, int], expectation: Any) -> None:
+def test_exception_handling_of_color_hsl(hsl: tuple[float, float, float], expectation: Any) -> None:
     with expectation:
         _ = ColorHSL(*hsl) is not None
 
 
 @pytest.mark.parametrize("hsl, expectation", MOCK_DATASET)
-def test_exception_handling_of_background_color_hsl(hsl: tuple[int, int, int], expectation: Any) -> None:
+def test_exception_handling_of_background_color_hsl(hsl: tuple[float, float, float], expectation: Any) -> None:
     with expectation:
         _ = BgColorHSL(*hsl) is not None
